bfdd: Add `--localaddr` option
Add an option to define a set of local addresses to listen on instead of
default binding to 0.0.0.0/::. Currently this option affects only
shop/mhop modes and not the echo mode due to differences in
implementation (including `IP_PACKET` socket) and because echo sockets
are not open by default. The behaviour is made simmilar to the `-l`
option for bgpd, namely in the way of handling several VRFs.