doc: Describe bgpd bind (`-l`) behaviour with VRFs
The `bgp_socket()` is called for each address in the `bm->addresses`
list (`-l`). Inside it an attempt is made to open and bind a socket in
the given VRF. If socket creation fails, bgpd will terminate unless we
are trying to open it in VRF (`BGP_INSTANCE_TYPE_VRF`).

Considering the case where there's address A in the default VRF and
address B in VRF1000, listening address A works, and B does not.

Signed-off-by: Mikhail Sokolovskiy <sokolmish@gmail.com>